+++ docs/html/search/all_1.js
@@ -1,5 +1,8 @@
 var searchData=
 [
-  ['bcdtodec_2',['bcdToDec',['../arithmetic_8h.html#a8547966ab0e711d97251ab6cb074e181',1,'ace_common']]],
-  ['buf_5f_3',['buf_',['../classace__common_1_1PrintStrBase.html#a0b8ea388f529134fdfc0cde78d866b9a',1,'ace_common::PrintStrBase']]]
+  ['bcdtodec_3',['bcdToDec',['../arithmetic_8h.html#a8547966ab0e711d97251ab6cb074e181',1,'ace_common']]],
+  ['binarysearch_4',['binarySearch',['../binarySearch_8h.html#abe3003dd71c6e85b71e7e7f0456c1acb',1,'ace_common']]],
+  ['binarysearch_2eh_5',['binarySearch.h',['../binarySearch_8h.html',1,'']]],
+  ['binarysearchbykey_6',['binarySearchByKey',['../binarySearch_8h.html#a284143acfd437ee8867a338213ed0044',1,'ace_common']]],
+  ['buf_5f_7',['buf_',['../classace__common_1_1PrintStrBase.html#a0b8ea388f529134fdfc0cde78d866b9a',1,'ace_common::PrintStrBase']]]
 ];