+++ docs/html/search/functions_0.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['bcdtodec_64',['bcdToDec',['../arithmetic_8h.html#a8547966ab0e711d97251ab6cb074e181',1,'ace_common']]]
+  ['bcdtodec_69',['bcdToDec',['../arithmetic_8h.html#a8547966ab0e711d97251ab6cb074e181',1,'ace_common']]],
+  ['binarysearch_70',['binarySearch',['../binarySearch_8h.html#abe3003dd71c6e85b71e7e7f0456c1acb',1,'ace_common']]],
+  ['binarysearchbykey_71',['binarySearchByKey',['../binarySearch_8h.html#a284143acfd437ee8867a338213ed0044',1,'ace_common']]]
 ];