+++ docs/html/search/functions_2.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['dectobcd_66',['decToBcd',['../arithmetic_8h.html#a3c833892a295cb0a698b974c18c8cdd8',1,'ace_common']]]
+  ['dectobcd_73',['decToBcd',['../arithmetic_8h.html#a3c833892a295cb0a698b974c18c8cdd8',1,'ace_common']]]
 ];