+++ src/AceCommon.h
@@ -36,6 +36,15 @@ SOFTWARE.
 #ifndef ACE_COMMON_H
 #define ACE_COMMON_H
 
+// Blacklist boards using new Arduino API due to incompatibilities. This
+// currently includes all megaAVR boards and SAMD21 boards using arduino::samd
+// >= 1.8.10. Boards using arduino:samd <= 1.8.9 or SparkFun:samd are fine.
+#if defined(ARDUINO_ARCH_MEGAAVR)
+#error MegaAVR not supported, https://github.com/bxparks/AceCommon/issues/8
+#elif defined(ARDUINO_ARCH_SAMD) && defined(ARDUINO_API_VERSION)
+#error SAMD21 with arduino:samd >= 1.8.10 not supported, https://github.com/bxparks/AceCommon/issues/9
+#endif
+
 #include "arithmetic/arithmetic.h"
 #include "pstrings/pstrings.h"
 #include "print_str/PrintStr.h"
@@ -46,9 +55,10 @@ SOFTWARE.
 #include "fstrings/FCString.h"
 #include "hash/djb2.h"
 #include "kstrings/KString.h"
+#include "algorithms/binarySearch.h"
 
 // Version format: "xx.yy.zz" => xxyyzz (without leading 0)
-#define ACE_COMMON_VERSION 10301
-#define ACE_COMMON_VERSION_STRING "1.3.1"
+#define ACE_COMMON_VERSION 10400
+#define ACE_COMMON_VERSION_STRING "1.4"
 
 #endif